impl. common Worker trait for running Workloads
